v1.3

+++ src/main/java/com/integrador/tp/model/Academico.java
@@ -19,7 +19,7 @@ public class Academico implements Serializable {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private float idAcademico;
+    private Long idAcademico;
     //-------------------------
            
     private String escuela;
