+++ src/main/java/com/integrador/tp/repository/AcademicoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AcademicoRepository extends JpaRepository <Academico, Float>{
+public interface AcademicoRepository extends JpaRepository <Academico, Long>{
     
 }
