+++ src/main/java/com/integrador/tp/service/IExperienciaService.java
@@ -5,7 +5,7 @@ import com.integrador.tp.model.Experiencia;
 public interface IExperienciaService {
      public void crearExperiencia(Experiencia exp);
     
-    public void borrarExperiencia(float id);
+    public void borrarExperiencia(Long id);
     
-    public Experiencia bucasrExperiencia(float id);
+    public Experiencia bucasrExperiencia(Long id);
 }
