+++ src/main/java/com/integrador/tp/service/ISeccionService.java
@@ -5,8 +5,8 @@ import com.integrador.tp.model.Seccion;
 public interface ISeccionService {
     public void crearSeccion (Seccion sec);
     
-    public void borrarSeccion (float id);
+    public void borrarSeccion (Long id);
     
-    public Seccion buscarSeccion(float id);
+    public Seccion buscarSeccion(Long id);
     
 }
