+++ src/main/java/com/integrador/tp/service/ISkillService.java
@@ -6,7 +6,7 @@ public interface ISkillService {
     
     public void crearSkill (Skill skill);
     
-    public void borrarSkill (float id);
+    public void borrarSkill (Long id);
     
-    public Skill buscarSkill (float id);
+    public Skill buscarSkill (Long id);
 }
